backend/kauri_ocr_service/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Message Queue
pika==1.3.2
celery==5.3.6

# OCR Engines
paddleocr==2.7.3
paddlepaddle==2.6.0
# surya-ocr==0.4.0  # Alternative OCR
# transformers==4.36.0  # For Qwen2.5-VL in production

# Image Processing
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pillow==10.2.0
pdf2image==1.17.0
pypdf2==3.0.1
img2pdf==0.5.1

# ML & NLP
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
spacy==3.7.2
# python -m spacy download fr_core_news_md

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.2
beautifulsoup4==4.12.3

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1
python-magic==0.4.27

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Storage
minio==7.2.3
boto3==1.34.34

# Validation & Data Processing
jsonschema==4.21.0
pycountry==23.12.11
babel==2.14.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.5.1

# Development
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
